+++ hatanaka/__init__.py
@@ -1,5 +1,5 @@
 from .general_compression import *
 from .hatanaka import *
 
-__version__ = '2.5.0'
+__version__ = '2.6.0'
 rnxcmp_version = '4.0.8'